addrbook

+++ jwbook/src/main/webapp/address_book/addrDelete.jsp
@@ -3,8 +3,11 @@
 <jsp:useBean id="abDAO" class="com.dao.AddrBookDAO" scope="application"/>    
 <%
 	String uname = request.getParameter("username");
-
-	abDAO.delete(uname);  //dao의 delete() 메소드 호출
 	
-	response.sendRedirect("addrList.jsp");  //삭제 후 목록 보기 페이지로 이동
+	if(session.getAttribute("userName") != null){
+		abDAO.delete(uname);  //dao의 delete() 메소드 호출
+		response.sendRedirect("./addrList.jsp");  //삭제 후 목록 보기 페이지로 이동
+	}else{
+		response.sendRedirect("./addrForm.jsp");  //주소 등록 페이지로 이동
+	}
 %>
